Sun animations
The sun is no longer stuck to the cursor, it trails behind it, short movements of the mouse result in slow movements of the sun, longer mouse movements cause the sun to follow the cursor more aggressively. I have no idea why this happens but it is kind of cool so I'll just leave it like that.