Umbrella now stops rain
But holy shit does this perform like garbage. It starts out fine but for some reason gets more and more out of sync... yeah, this is a problem, I will definitely need to investiate this. It works the first 10 seconds, though and that's the most important thing about a demo like this.