Updated backend logic for obtaining weather
The sidebar item now changes background color depending on the conditions. These conditions need to be set by setting the weather, otherwise the binding engine won't update the UI. Also minor layout refiniements, which means the sun to the left side of this control is now less thicc... well, wide.