Adjusted target framework again
We now have the minimum requirement of Windows 11 22H2 to match the one of the Pikouna Engine itself, which should help with importing easily. We have no plans to support Windows 10 at the moment, as it will be out of support when the app launches, most likely, and as we do not have any machines to test on.