Respecting "no animations setting" and adjusting animations
The sidebar now respects the setting in Windows Settings - Accessibility - Visual effects - Animation effects, it does not animate when this setting is disabled. Also tweaked animations a little - the ripple is a bit less impactful, as it now drops off much sooner, and the curve for the sidebar entrance has been slightly adjusted.