Made content cards translucent
The weather content widgets are now using a blur effect.
Also updated the sidebar animation a little to no longer entrance with a BackEase because at large amplitudes it looks awkward and at low amplitudes it didn't look like it was intended but rather like it was a mistake and it's awkwardly jumping somewhere. We use CircleEase for both transitions now.