ViewModel and sidebar changes
The sidebar now displays the highest and lowest apparent temperature and they are truncated to only the highest and in the worst case removed if the weather description doesn't fit - worst case is that the weather description gets a flyout on hover.
Also made sure to include the location in the widgets and in the ViewModel to display it in the body of the app later.