Alignment changes
The sidebar and main weather panel received alignment changes to comply with Microsoft's rule of scaling alignments to multiples of four and to comply with Microsoft's recommendation of choosing a window border of 12px, rather than the previous 5px all around, we now use 12px on the sidebar (on either side) and 4px around the weather view container - the sidebar items also get a spacing of 8px. This is still subject to change of course, since I am not entirely happy with how it currently looks.